refactor(shared): share offline action types between web and mobile
- Create @volleykit/shared/offline module with shared action types
- Move CompensationUpdateData, ActionStatus, and action interfaces to shared
- Update web-app to re-export from shared
- Update mobile to re-export from shared with mobile-specific subset
- Simplify mobile action-sync by using shared type field names
