+++ packages/shared/src/offline/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Shared offline mode types for web and mobile apps.
+ */
+
+export * from './action-types'
